create data_struc.h
