+++ src/exec/here_doc.c
@@ -6,12 +6,11 @@
 /*   By: abasdere <abasdere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/09 11:09:35 by andymalgonn       #+#    #+#             */
-/*   Updated: 2025/01/30 14:22:03 by abasdere         ###   ########.fr       */
+/*   Updated: 2025/01/31 14:20:19 by abasdere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-#include "exec.h"
 
 char	*find_heredoc_file(void)
 {
