+++ src/tokens/string_tokenizing.c
@@ -3,15 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   string_tokenizing.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gmoulin <gmoulin@student.42.fr>            +#+  +:+       +#+        */
+/*   By: abasdere <abasdere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 17:33:45 by gmoulin           #+#    #+#             */
-/*   Updated: 2025/01/30 16:19:59 by gmoulin          ###   ########.fr       */
+/*   Updated: 2025/01/31 14:21:49 by abasdere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-#include "tokens.h"
 
 int	tokenize_s_quote(char **rl, t_token **head)
 {
